Generalize matrix with weights

+++ src/main/java/ru/itmo/ctlab/hict/hict_library/visualization/RawTileWithWeights.java
@@ -27,6 +27,6 @@ package ru.itmo.ctlab.hict.hict_library.visualization;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-public record RawTileWithWeights(long @NotNull [][] values, double @Nullable [] rowWeights,
+public record RawTileWithWeights(double @NotNull [][] values, double @Nullable [] rowWeights,
                                  double @Nullable [] columnWeights) {
 }
